refactoring and package name add initial fastlane stuff
