+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdkVersion 30
     buildToolsVersion "29.0.2"
     defaultConfig {
-        applicationId "org.chickenhook.AndroidHackingTemplate"
+        applicationId "org.chickenhook.parcelexploittester"
         minSdkVersion 19
         targetSdkVersion 31
         versionCode 1
