+++ app/src/main/java/org/chickenhook/parcelexploittester/MainActivity.kt
@@ -1,12 +1,13 @@
-package org.chickenhook.AndroidHackingTemplate
+package org.chickenhook.parcelexploittester
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import android.widget.Button
-import org.chickenhook.AndroidHackingTemplate.penetration.ClipboardManagerCrash
-import org.chickenhook.AndroidHackingTemplate.penetration.PackageManagerCrash
-import org.chickenhook.AndroidHackingTemplate.penetration.ServiceManagerCrash
+import org.chickenhook.parcelexploittester.R
+import org.chickenhook.parcelexploittester.penetration.ClipboardManagerCrash
+import org.chickenhook.parcelexploittester.penetration.PackageManagerCrash
+import org.chickenhook.parcelexploittester.penetration.ServiceManagerCrash
 
 //import org.chickenhook.myandroidhackinglibrary.NativeInterface
