+++ app/src/main/java/org/chickenhook/parcelexploittester/penetration/ClipboardManagerCrash.kt
@@ -1,4 +1,4 @@
-package org.chickenhook.AndroidHackingTemplate.penetration
+package org.chickenhook.parcelexploittester.penetration
 
 import android.annotation.SuppressLint
 import android.app.Activity
