+++ app/src/main/java/org/chickenhook/parcelexploittester/penetration/IExploit.java
@@ -1,4 +1,4 @@
-package org.chickenhook.AndroidHackingTemplate.penetration;
+package org.chickenhook.parcelexploittester.penetration;
 
 import android.content.Context;
