+++ app/src/main/java/org/chickenhook/parcelexploittester/penetration/PackageManagerCrash.kt
@@ -1,4 +1,4 @@
-package org.chickenhook.AndroidHackingTemplate.penetration
+package org.chickenhook.parcelexploittester.penetration
 
 import android.content.Context
 import android.os.Build
