+++ fastlane/metadata/android/en-US/changelogs/2.txt
@@ -1,0 +1,1 @@
+* initial release
