+++ fastlane/metadata/android/en-US/short_description.txt
@@ -1,0 +1,1 @@
+An App intended for testing against the Android Parcel/Binder framework. Check if your device is affected NOW!
