+++ fastlane/metadata/android/en-US/title.txt
@@ -1,0 +1,1 @@
+Parcel.java Exploit - Pentest the Android Parcel/Binder protocol
