+++ fastlane/metadata/android/en-US/video.txt
@@ -1,0 +1,1 @@
+https://www.youtube.com/watch?v=-_be1LBz7eA&t=9s
