--- myAndroidHackingLibrary/src/main/AndroidManifest.xml
@@ -1,2 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.chickenhook.myandroidhackinglibrary" />
