--- myAndroidHackingLibrary/src/main/java/org/chickenhook/myandroidhackinglibrary/NativeInterface.java
@@ -1,9 +1,0 @@
-package org.chickenhook.myandroidhackinglibrary;
-
-public class NativeInterface {
-    static {
-        System.loadLibrary("myNativeLib");
-    }
-
-    public static native void installHooks();
-}
